requirements.txt
crewai
google-generativeai
python-dotenv
streamlit
langchain-google-genai==2.0.10
langchain-core==0.3.75
